+++ examples/data-sources/meshstack_workspace/data-source.tf
@@ -1,0 +1,5 @@
+data "meshstack_workspace" "example" {
+  metadata = {
+    name = "my-workspace-identifier"
+  }
+}