+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## v0.7.2
+## v0.8.0
+
+FEATURES:
+- Added `meshstack_workspace` resource.
+- Added `meshstack_workspace` data source.
 
 FIXES:
 - Allow `value_code` in `meshstack_building_block_v2` and `meshstack_building_block` resources.